db-service/pyproject.toml
[project]
name = "db-service"
version = "0.1.0"
description = "Это микросервис, который имплементирует бизнес-логику взаимодействия с СУБД агрегатора новостей."
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.5",
    "django-cache-memoize>=0.2.1",
    "django-elasticsearch-dsl==8.2",
    "django-elasticsearch-dsl-drf-alt==0.22.9",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "elasticsearch==8.18.0",
    "elasticsearch-dsl==8.18.0",
    "ipython>=9.5.0",
    "kombu>=5.5.4",
    "pika>=1.3.2",
    "pika-stubs>=0.1.3",
    "psycopg[binary]>=3.2.9",
    "redis>=6.4.0",
    "ruff>=0.12.8",
    "setuptools>=80.9.0",
]


[tool.ruff]
fix = true

exclude = [
    ".git",
    ".venv",
    "migrations",
    "__pycache__",
]

# Выбор наборов правил
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort - сортировка импортов
    "UP",   # pyupgrage
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "RUF",  # Ruff-specific rules
]

# Игнорируем некоторые ошибки
ignore = [
    "E501",   # Длина строки превышает лимит
    "E401",   # Импорт не используется
    "B008",   # do not perform funcion call in argument defaults
]

target-version = "py313"
line-length = 88

[tool.ruff.isort]
known-first-party = ["db-service"]

[tool.ruff.per-file-ignores]

[tool.ruff.format]
quote-style = "double"  # какие кавычки использовать
indent-style = "space"  # форматирование отступов
line-ending = "auto"
